backend/requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
